refactor: split tween.js into separate files and bundle it with rollup.
