+++ .gitignore
@@ -2,3 +2,4 @@
 build/npm*
 node_modules
 npm-debug.log
+dist
