+++ package.json
@@ -2,7 +2,7 @@
   "name": "tween.js",
   "description": "Super simple, fast and easy to use tweening engine which incorporates optimised Robert Penner's equations.",
   "version": "0.0.0-development",
-  "main": "src/Tween.js",
+  "main": "dist/TWEEN.js",
   "homepage": "https://github.com/tweenjs/tween.js",
   "repository": {
     "type": "git",
@@ -18,10 +18,13 @@
   ],
   "dependencies": {},
   "scripts": {
+    "pretest": "npm run build",
     "test": "npm run test-unit && npm run test-correctness && npm run test-style",
     "test-unit": "nodeunit test/unit/nodeunitheadless.js",
-    "test-correctness": "jshint --config test/jshintrc src/Tween.js",
-    "test-style": "jscs --config test/jscs.json src/Tween.js",
+    "test-correctness": "jshint --config test/jshintrc src",
+    "test-style": "jscs --config test/jscs.json src",
+    "build": "rollup --config",
+    "prepublish": "npm run test",
     "semantic-release": "semantic-release pre && npm publish && semantic-release post"
   },
   "author": "tween.js contributors (https://github.com/tweenjs/tween.js/graphs/contributors)",
@@ -29,6 +32,8 @@
     "jscs": "^2.2.0",
     "jshint": "^2.8.0",
     "nodeunit": "^0.9.1",
+    "rollup": "^0.41.6",
+    "rollup-plugin-buble": "^0.15.0",
     "semantic-release": "^6.3.2"
   }
 }
