+++ test/unit/nodeunitheadless.js
@@ -1,5 +1,5 @@
 // We will unit test both the original library and the minified version
-var TWEEN_uncompressed = require('../../src/Tween.js');
+var TWEEN_uncompressed = require('../../dist/TWEEN.js');
 var getTests = require('./tests');
 
 module.exports = {
